Move CustomVizEmptyState to EE
