+++ frontend/src/metabase/plugins/oss/custom-viz.ts
@@ -21,6 +21,7 @@ const getDefaultPluginCustomViz = () => ({
   ManageCustomVizPage: PluginPlaceholder as ComponentType<any>,
   CustomVizPage: PluginPlaceholder as ComponentType<any>,
   CustomVizDevPage: PluginPlaceholder as ComponentType<any>,
+  CustomVizEmptyState: PluginPlaceholder as ComponentType<any>,
 
   // Hooks & functions
   useAutoLoadCustomVizPlugin: (_display: string | undefined) => ({
